pyproject.toml
[project]
name = "tomaba"
version = "0.1.0"
description = ""
authors = [
    {name = "Matt Mahowald",email = "mahowaldmatt@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.1",
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "mistralai (>=1.5.0,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
